Fix Gorouter property tests
This commit deduplicates and fixes tests added as part of allowing
Gorouter to talk to NATS over mTLS.

Huge thanks to @ameowlia for figuring out how to fix the tests.
https://github.com/cloudfoundry/routing-release/pull/204#issuecomment-875068923
