+++ jobs/gorouter/templates/gorouter.yml.erb
@@ -9,7 +9,7 @@ def property_or_link(description, property, link_path, link_name=nil)
       return link_object.p(link_path)
     end
   end
-  raise RuntimeError, "#{description} not found in properties nor in \"#{link_name}\" link."
+  raise RuntimeError, "#{description} not found in properties nor in \"#{link_name}\" link. This value can be specified using the \"#{property}\" property."
 end
 %>
