+++ spec/gorouter_templates_spec.rb
@@ -5,6 +5,7 @@ require 'yaml'
 require 'json'
 require 'bosh/template/test'
 require 'bosh/template/evaluation_context'
+require 'spec_helper'
 
 TEST_CERT = 'some
 
@@ -548,136 +549,11 @@ describe 'gorouter' do
             deployment_manifest_fragment['routing_api']['enabled'] = true
           end
 
-          class LinkConfiguration
-            attr_reader :description, :property, :link, :parsed_yaml_property
-
-            def initialize(description:, property:, link:, parsed_yaml_property:)
-              @description = description
-              @property = property
-              @link = link
-              @parsed_yaml_property = parsed_yaml_property
-            end
-
-            def link_namespace
-              link.split('.').first
-            end
-          end
-
-          shared_examples 'overridable_link' do |link_config|
-            def get_at_property(hash, property)
-              property_chain = property.split('.')
-
-              get_this = hash
-              property_chain.each do |getter|
-                get_this = get_this.fetch(getter)
-              end
-
-              get_this
-            end
-
-            def set_at_property(hash, property, value)
-              property_chain = property.split('.')
-
-              getters = property_chain[0..-2]
-              setter = property_chain.last
-
-              set_this = hash
-              getters.each do |getter|
-                set_this = set_this.fetch(getter)
-              end
-
-              set_this.store(setter, value)
-            end
-
-            def delete_at_property(hash, property)
-              property_chain = property.split('.')
-
-              getters = property_chain[0..-2]
-              setter = property_chain.last
-
-              set_this = hash
-              getters.each do |getter|
-                set_this = set_this.fetch(getter)
-              end
-
-              set_this.delete(setter)
-            end
-
-            context 'when the link is not provided' do
-              context 'when the property is set' do
-                before do
-                  # TODO: constant it
-                  set_at_property(deployment_manifest_fragment, link_config.property, property_value)
-                end
-
-                it 'should prefer the value in the properties' do
-                  expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(property_value)
-                end
-              end
-
-              context 'when the property is not set' do
-                before do
-                  delete_at_property(deployment_manifest_fragment, link_config.property)
-                end
-
-                it 'should error' do
-                  expect do
-                    parsed_yaml
-                  end.to raise_error(
-                    RuntimeError,
-                    "#{link_config.description} not found in properties nor in \"#{link_config.link_namespace}\" link. This value can be specified using the \"#{link_config.property}\" property."
-                  )
-                end
-              end
-            end
-
-            context 'when the link is provided' do
-              def make_properties(link, link_property)
-                property_chain = link.split('.').reverse
-
-                property_chain.reduce(link_property) do |memo, obj|
-                  { obj => memo }
-                end
-              end
-
-              let(:links) do
-                [
-                  Bosh::Template::Test::Link.new(
-                    name: link_config.link_namespace,
-                    properties: make_properties(link_config.link, link_value)
-                  )
-                ]
-              end
-
-              let(:rendered_template) { template.render(deployment_manifest_fragment, consumes: links) }
-
-              context 'when the property is set' do
-                before do
-                  set_at_property(deployment_manifest_fragment, link_config.property, property_value)
-                end
-
-                it 'should prefer the value in the properties' do
-                  expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(property_value)
-                end
-              end
-
-              context 'when the property is not set' do
-                before do
-                  delete_at_property(deployment_manifest_fragment, link_config.property)
-                end
-
-                it 'should render the value from the link' do
-                  expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(link_value)
-                end
-              end
-            end
-          end
-
           describe 'routing API port' do
             it_behaves_like 'overridable_link', LinkConfiguration.new(
               description: 'Routing API port',
               property: 'routing_api.port',
-              link: 'routing_api.mtls_port',
+              link_path: 'routing_api.mtls_port',
               parsed_yaml_property: 'routing_api.port'
             )
           end
@@ -722,7 +598,7 @@ describe 'gorouter' do
             it_behaves_like 'overridable_link', LinkConfiguration.new(
               description: 'Routing API server CA certificate',
               property: 'routing_api.ca_certs',
-              link: 'routing_api.mtls_ca',
+              link_path: 'routing_api.mtls_ca',
               parsed_yaml_property: 'routing_api.ca_certs'
             )
           end
@@ -743,7 +619,7 @@ describe 'gorouter' do
             it_behaves_like 'overridable_link', LinkConfiguration.new(
               description: 'Routing API client private key',
               property: 'routing_api.private_key',
-              link: 'routing_api.mtls_client_key',
+              link_path: 'routing_api.mtls_client_key',
               parsed_yaml_property: 'routing_api.private_key'
             )
           end
@@ -774,7 +650,7 @@ describe 'gorouter' do
             it_behaves_like 'overridable_link', LinkConfiguration.new(
               description: 'Routing API client certificate',
               property: 'routing_api.cert_chain',
-              link: 'routing_api.mtls_client_cert',
+              link_path: 'routing_api.mtls_client_cert',
               parsed_yaml_property: 'routing_api.cert_chain'
             )
           end
@@ -785,139 +661,12 @@ describe 'gorouter' do
         let(:property_value) { ('a'..'z').to_a.shuffle.join }
         let(:link_value) { ('a'..'z').to_a.shuffle.join }
 
-        class LinkConfiguration
-          attr_reader :description, :property, :link, :parsed_yaml_property
-
-          def initialize(description:, property:, link:, parsed_yaml_property:)
-            @description = description
-            @property = property
-            @link = link
-            @parsed_yaml_property = parsed_yaml_property
-          end
-
-          def link_namespace
-            link.split('.').first
-          end
-        end
-
-        shared_examples 'overridable_link' do |link_config|
-          def get_at_property(hash, property)
-            property_chain = property.split('.')
-
-            get_this = hash
-            property_chain.each do |getter|
-              getter = getter.to_i if get_this.class == Array
-              get_this = get_this.fetch(getter)
-            end
-
-            get_this
-          end
-
-          def set_at_property(hash, property, value)
-            property_chain = property.split('.')
-
-            getters = property_chain[0..-2]
-            setter = property_chain.last
-
-            set_this = hash
-            getters.each do |getter|
-              getter = getter.to_i if set_this.class == Array
-              set_this = set_this.fetch(getter)
-            end
-
-            set_this.store(setter, value)
-          end
-
-          def delete_at_property(hash, property)
-            property_chain = property.split('.')
-
-            getters = property_chain[0..-2]
-            setter = property_chain.last
-
-            set_this = hash
-            getters.each do |getter|
-              getter = getter.to_i if set_this.class == Array
-              set_this = set_this.fetch(getter)
-            end
-
-            set_this.delete(setter)
-          end
-
-          context 'when the link is not provided' do
-            context 'when the property is set' do
-              before do
-                # TODO: constant it
-                set_at_property(deployment_manifest_fragment, link_config.property, property_value)
-              end
-
-              it 'should prefer the value in the properties' do
-                expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(property_value)
-              end
-            end
-
-            context 'when the property is not set' do
-              before do
-                delete_at_property(deployment_manifest_fragment, link_config.property)
-              end
-
-              it 'should error' do
-                expect do
-                  parsed_yaml
-                end.to raise_error(
-                  RuntimeError,
-                  "#{link_config.description} not found in properties nor in \"#{link_config.link_namespace}\" link. This value can be specified using the \"#{link_config.property}\" property."
-                )
-              end
-            end
-          end
-
-          context 'when the link is provided' do
-            def make_properties(link, link_property)
-              property_chain = link.split('.').reverse
-
-              property_chain.reduce(link_property) do |memo, obj|
-                { obj => memo }
-              end
-            end
-
-            let(:links) do
-              [
-                Bosh::Template::Test::Link.new(
-                  name: link_config.link_namespace,
-                  properties: make_properties(link_config.link, link_value)
-                )
-              ]
-            end
-
-            let(:rendered_template) { template.render(deployment_manifest_fragment, consumes: links) }
-
-            context 'when the property is set' do
-              before do
-                set_at_property(deployment_manifest_fragment, link_config.property, property_value)
-              end
-
-              it 'should prefer the value in the properties' do
-                expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(property_value)
-              end
-            end
-
-            context 'when the property is not set' do
-              before do
-                delete_at_property(deployment_manifest_fragment, link_config.property)
-              end
-
-              it 'should render the value from the link' do
-                expect(get_at_property(parsed_yaml, link_config.parsed_yaml_property)).to eq(link_value)
-              end
-            end
-          end
-        end
-
         describe 'NATS port' do
           it_behaves_like 'overridable_link', LinkConfiguration.new(
             description: 'NATS server port number',
             property: 'nats.port',
-            link: 'nats-tls.nats.port',
+            link_path: 'nats.port',
+            link_namespace: 'nats-tls',
             parsed_yaml_property: 'nats.hosts.0.port'
           )
         end
@@ -962,7 +711,8 @@ describe 'gorouter' do
           it_behaves_like 'overridable_link', LinkConfiguration.new(
             description: 'NATS server CA certificate',
             property: 'nats.ca_certs',
-            link: 'nats-tls.nats.external.tls.ca',
+            link_path: 'nats.external.tls.ca',
+            link_namespace: 'nats-tls',
             parsed_yaml_property: 'nats.ca_certs'
           )
         end
